docs/examples/use_cases/tensorflow/efficientdet/requirements.txt
argparse-utils
absl-py
tensorflow-gpu
